Added Members Page
Added a members page where a list of members on a workspace pop up with a button that has 3 functionalities. Right now, there is an error when it comes to the members page with displaying the role of the member, the error doesn't stop any functionalities but it is still something to look into.