Projects Settings (Unfinished)
Created a page for projects as well as a settings component to it. However, there is an error in fetching the projects so it sends you to the error page. Another error in with the submit projects form where I tried to send the user to the url of the project but it is giving me an error.